refactor subscriber service, added DeactiveSubscriber service
